+++ tests/test_docs_hooks.py
@@ -117,7 +117,7 @@ def test_url(job: JobBase) -> None:
     # noinspection PyTypeChecker
     with JobState(None, job) as job_state:  # type: ignore[arg-type]
         data, mime_type = FilterBase.auto_process(job_state, data, '')
-        for filter_kind, subfilter in FilterBase.normalize_filter_list(job_state.job.filter):
+        for filter_kind, subfilter in FilterBase.normalize_filter_list(job_state.job.filters):
             data, mime_type = FilterBase.process(filter_kind, subfilter, job_state, data, '')
 
         expected_output_data = d['output']